feat(raster): return RasterResult from WMTS/WCS providers
Providers now return RasterResult (image + bbox) instead of empty/
metadata GeoDataFrames. OutputManager saves GeoTIFF + JPEG per layer.
RecipeRunner.execute() returns (vector_layers, raster_layers) tuple.
GLBExporter gains add_raster_plane() for textured ground planes in GLB.

+++ giskit/protocols/wmts.py
@@ -262,7 +262,7 @@ class WMTSProtocol(Protocol):
         self,
         bbox: tuple[float, float, float, float],
         product: str,
-        resolution: int,
+        resolution: float,
         crs: str = "EPSG:28992",
         zoom: Optional[int] = None,
         progress_callback: Optional[Any] = None,
